refactor: extract handleRequires() to file
