+++ lib/cli/run.js
@@ -16,12 +16,8 @@ const {
   createMissingArgumentError
 } = require('../errors');
 
-const {
-  list,
-  handleRequires,
-  validateLegacyPlugin,
-  runMocha
-} = require('./run-helpers');
+const {list, validateLegacyPlugin, runMocha} = require('./run-helpers');
+const {handleRequires} = require('./handle-requires');
 const {ONE_AND_DONES, ONE_AND_DONE_ARGS} = require('./one-and-dones');
 const debug = require('debug')('mocha:cli:run');
 const defaults = require('../mocharc');
