+++ lib/nodejs/worker.js
@@ -12,7 +12,8 @@ const {
 } = require('../errors');
 const workerpool = require('workerpool');
 const Mocha = require('../mocha');
-const {handleRequires, validateLegacyPlugin} = require('../cli/run-helpers');
+const {validateLegacyPlugin} = require('../cli/run-helpers');
+const {handleRequires} = require('../cli/handle-requires');
 const d = require('debug');
 const debug = d.debug(`mocha:parallel:worker:${process.pid}`);
 const isDebugEnabled = d.enabled(`mocha:parallel:worker:${process.pid}`);
